+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Last Changes
 
+- [#6](https://github.com/LaxarJS/laxar-tooling/issues/6): support preprocessed theme-, layout-, widget- and control-artifacts
+  + **BREAKING CHANGE:** see ticket for details
 - [#5](https://github.com/LaxarJS/laxar-tooling/issues/5): implement new path resolution scheme
   + **BREAKING CHANGE:** see ticket for details
 - [#4](https://github.com/LaxarJS/laxar-tooling/issues/4): converted the source code to ES2015, pre-built with babel